Refactor fretboardUtils
