+++ src/utils/fretboard/containsSharpOrFlat.ts
@@ -1,0 +1,6 @@
+export function containsSharpOrFlat(note: string) {
+  if (note.includes("♭") || note.includes("♯")) {
+    return true
+  }
+  return false
+}
