+++ src/utils/fretboard/getNote.ts
@@ -1,0 +1,79 @@
+import { AccidentalMode } from "src/apps/settings/settingsState"
+import { containsSharpOrFlat } from "./containsSharpOrFlat"
+import { getString } from "./getString"
+import { isEmpty, random } from "lodash"
+import { StringFocus } from "src/apps/notes/state/notesSettingsState"
+
+import {
+  GuitarString,
+  NotePosition,
+  fretboardNoteMap,
+  IntervalLabels,
+} from "../types"
+
+/**
+ * Returns a note from the notes array.
+ *
+ * const note = getNote({
+ *   mode: 'flats',
+ *   position: [1, 3] // first string, third fret
+ * })
+ */
+
+export interface Note {
+  note: string
+  string?: GuitarString
+  position: NotePosition
+  interval?: IntervalLabels
+}
+
+export function getNote(
+  props: {
+    accidentalMode?: AccidentalMode
+    position?: NotePosition
+    startingFret?: number
+    stringFocus?: StringFocus
+  } = {}
+): Note {
+  const {
+    accidentalMode = "flats",
+    position,
+    startingFret = 1,
+    stringFocus = 0,
+  } = props
+
+  let string
+  let note
+
+  if (!position) {
+    string = stringFocus || random(1, 6)
+    note = random(startingFret, 12)
+  } else {
+    string = position[0]
+    note = position[1]
+  }
+
+  // Use a 1-based index to follow guitar idioms. We don't subtract from `noteId`
+  // because 0 (as in [6, 0]) refers to an open string -- in this case the low `E`.
+  const stringName = getString(string)
+  string--
+
+  const noteName = fretboardNoteMap[accidentalMode][string][note]
+
+  // Re-run function if we return an invalid result
+  const showAccidentals = accidentalMode !== "naturals"
+  if (
+    (!showAccidentals && containsSharpOrFlat(noteName)) ||
+    isEmpty(noteName)
+  ) {
+    return getNote(props)
+  }
+
+  const foundNote: Note = {
+    note: noteName,
+    string: stringName,
+    position: [string, note],
+  }
+
+  return foundNote
+}
