+++ src/utils/types.ts
@@ -1,0 +1,87 @@
+export type StringRange = 1 | 2 | 3 | 4 | 5 | 6
+
+export type NoteRange = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12 | 13 // prettier-ignore
+
+export type GuitarString = "E" | "b" | "g" | "d" | "a" | "e"
+
+export type NotePosition = [StringRange, NoteRange]
+
+export type IntervalLabels =
+  | "1"
+  | "unison"
+  | "minor 2nd"
+  | "♭2"
+  | "major 2nd"
+  | "2"
+  | "minor 3rd"
+  | "♭3"
+  | "major 3rd"
+  | "3"
+  | "perfect 4th"
+  | "4"
+  | "#4/♭5"
+  | "♭5"
+  | "dim 5th"
+  | "aug 4th"
+  | "♭4"
+  | "#5"
+  | "perfect 5th"
+  | "5"
+  | "#5/♭6"
+  | "minor 6th"
+  | "♭6"
+  | "aug 5th"
+  | "major 6th"
+  | "6"
+  | "minor 7th"
+  | "♭7"
+  | "major 7th"
+  | "7"
+  | "octave"
+
+export const intervalList: Partial<IntervalLabels>[] = [
+  "1",
+  "♭2",
+  "2",
+  "♭3",
+  "3",
+  "4",
+  "#4/♭5",
+  "5",
+  "#5/♭6",
+  "6",
+  "♭7",
+  "7",
+  "1",
+]
+
+// TODO:
+// Surely this map can be done dynamically based upon empty slots in the
+// `naturals` array below.
+
+export const fretboardNoteMap = {
+  flats: [
+    ["E", "F", "G♭", "G", "A♭", "A", "B♭", "B", "C", "D♭", "D", "E♭", "E"],
+    ["B", "C", "D♭", "D", "E♭", "E", "F", "G♭", "G", "A♭", "A", "B♭", "B"],
+    ["G", "A♭", "A", "B♭", "B", "C", "D♭", "D", "E♭", "E", "F", "G♭", "G"],
+    ["D", "E♭", "E", "F", "G♭", "G", "A♭", "A", "B♭", "B", "C", "D♭", "D"],
+    ["A", "B♭", "B", "C", "D♭", "D", "E♭", "E", "F", "G♭", "G", "A♭", "A"],
+    ["E", "F", "G♭", "G", "A♭", "A", "B♭", "B", "C", "D♭", "D", "E♭", "E"],
+  ],
+  sharps: [
+    ["E", "F", "F♯", "G", "G♯", "A", "A♯", "B", "C", "C♯", "D", "D♯", "E"],
+    ["B", "C", "C♯", "D", "D♯", "E", "F", "F♯", "G", "G♯", "A", "A♯", "B"],
+    ["G", "G♯", "A", "B♯", "B", "C", "C♯", "D", "D♯", "E", "F", "G♯", "G"],
+    ["D", "D♯", "E", "F", "F♯", "G", "G♯", "A", "A♯", "B", "C", "C♯", "D"],
+    ["A", "A♯", "B", "C", "C♯", "D", "D♯", "E", "F", "F♯", "G", "G♯", "A"],
+    ["E", "F", "F♯", "G", "G♯", "A", "A♯", "B", "C", "C♯", "D", "D♯", "E"],
+  ],
+  naturals: [
+    ["E", "F", "", "G", "", "A", "", "B", "C", "", "D", "", "E"],
+    ["B", "C", "", "D", "", "E", "F", "", "G", "", "A", "", "B"],
+    ["G", "", "A", "", "B", "C", "", "D", "", "E", "F", "", "G"],
+    ["D", "", "E", "F", "", "G", "", "A", "", "B", "C", "", "D"],
+    ["A", "", "B", "C", "", "D", "", "E", "F", "", "G", "", "A"],
+    ["E", "F", "", "G", "", "A", "", "B", "C", "", "D", "", "E"],
+  ],
+}
